Add JsonClient which takes provider as arg
